+++ src/grt/src/fastroute/src/maze3D.cpp
@@ -666,7 +666,7 @@ void FastRouteCore::mazeRouteMSMDOrder3D(int expand,
     const int netID = tree_order_pv_[orderIndex].treeIndex;
 
     FrNet* net = nets_[netID];
-    int edge_cost = 0;
+    int8_t edge_cost = 0;
 
     int enlarge = expand;
     const int num_terminals = sttrees_[netID].num_terminals;
